+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * Update CI
 * Update to BlossomLib 2
+* Add /lastdeath command
 
 # 1.0.0